requirements.txt
streamlit>=1.30.0
sentence-transformers==2.7.0
faiss-cpu>=1.7.4
scikit-learn>=1.3.0
datasets>=2.14.0
pandas>=2.0.0
numpy>=1.24.0
torch==2.3.0
transformers==4.41.0
groq
python-dotenv
gTTS>=2.3.0
SpeechRecognition>=3.10.0
audio-recorder-streamlit>=0.0.8
pydub>=0.25.0
